Module icalendar::component

The component itself

Structs

Component
A fully featured and complete Component with only valid attributes

InvalidComponentError
Tells you that the given string is not a valid component

InvalidParameterError
Tells you that the given string is not a valid parameter

InvalidPropertyError
Tells you that the given string is not a valid property

Property
A valid property for it's parent Component

Enums

ComponentType
Taken from the RFC5545

ParameterType
Taken from RFC 5545

PropertyType
Taken from the RFC 5545